% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pbsmrtpipe.R
\name{dictToResolvedToolContract}
\alias{dictToResolvedToolContract}
\title{Convert a dict to a Resolved Task Contract}
\usage{
dictToResolvedToolContract(d)
}
\description{
Convert a dict to a Resolved Task Contract
}

