% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleTasks.R
\name{examplefastaReport}
\alias{examplefastaReport}
\title{Example Task for Testing with emitting a report}
\usage{
examplefastaReport(pathToFasta, report)
}
\description{
inputs = [FileTypes.Fasta]
outputs = [FileTypes.Report]
}

