% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleTasks.R
\name{examplefilterFastaTask}
\alias{examplefilterFastaTask}
\title{Example Task for testing
inputs = [FileTypes.Fasta]
outputs = [FileTypes.Fasta]}
\usage{
examplefilterFastaTask(pathToFasta, filteredFasta, minSequenceLength)
}
\description{
Example Task for testing
inputs = [FileTypes.Fasta]
outputs = [FileTypes.Fasta]
}

