% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exampleTasks.R
\name{exampleTask01}
\alias{exampleTask01}
\title{Example Task for testing
inputs = [FileTypes.Fasta]
outputs = [FileTypes.Fasta]}
\usage{
exampleTask01(pathToFasta, filteredFasta, minSequenceLength)
}
\description{
Example Task for testing
inputs = [FileTypes.Fasta]
outputs = [FileTypes.Fasta]
}

