% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exampleTasks.R
\name{exampleTask02}
\alias{exampleTask02}
\title{Example Task for Testing with emitting a report}
\usage{
exampleTask02(pathToFasta, report)
}
\description{
inputs = [FileTypes.Fasta]
outputs = [FileTypes.Report]
}

