% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\docType{data}
\name{FileTypes}
\alias{FileTypes}
\title{Registry for all the PacBio defined file types}
\format{An object of class \code{list} of length 6.}
\usage{
FileTypes
}
\description{
Registry for all the PacBio defined file types
}
\keyword{datasets}

