% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/driver.R
\name{getParser}
\alias{getParser}
\title{Fundamental argument parser to emit tool contracts and
Run tool contracts}
\usage{
getParser()
}
\description{
Fundamental argument parser to emit tool contracts and
Run tool contracts
}

