% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphicDefaults.R
\name{getPBColorScale}
\alias{getPBColorScale}
\title{Get the default color scheme, adding more colors if they are available.}
\usage{
getPBColorScale(numLevels = 9)
}
\description{
Get the default color scheme, adding more colors if they are available.
}

