% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphicDefaults.R
\name{getPBFillScale}
\alias{getPBFillScale}
\title{Get the default color scheme for fills, adding more colors if the current
palette is maxed out}
\usage{
getPBFillScale(numLevels = 9)
}
\description{
Get the default color scheme for fills, adding more colors if the current
palette is maxed out
}

