% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{loadResolvedToolContractFromPath}
\alias{loadResolvedToolContractFromPath}
\title{Load a Resolved Tool contract from json file}
\usage{
loadResolvedToolContractFromPath(path)
}
\description{
Load a Resolved Tool contract from json file
}

