% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver.R
\name{mainRegisteryMainArgs}
\alias{mainRegisteryMainArgs}
\title{CLI entry point. Parses CLI args and calls registery Runner}
\usage{
mainRegisteryMainArgs(registry)
}
\description{
CLI entry point. Parses CLI args and calls registery Runner
}

