% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar.R
\name{pbreporter}
\alias{pbreporter}
\title{Sugar object to make it easier to generate reports.}
\usage{
pbreporter(conditionFile, outputFile, reportid, version = "0.0.1")
}
\arguments{
\item{reportid}{A lowercase/no-special character ID.}

\item{version}{The version number (1.0.0)}

\item{reSeqConditions}{The file with the resequencing conditions.}

\item{reportOutputPath}{The path of the report JSON file to write.}
}
\description{
When writing a report, there is a lot of "state" to account for.  Plots and
tables must be written to the same output directory, and whenever a file is
output, we have to update the pbreport to ensure it shows up in the GUI.
}
\details{
Passing all this information around gets annoying fast. R is a functional
language and does not like keeping track of state very much. To work around
this, we are going to use the closure approach outlined in Section 5.4 of
"Software for Data Analysis" by John Chambers.  The basic idea is that we'll
define functions for saving plots/data in a closure that contains an
environment with the objects we want to update.

This could also be accomplished using the newer reference classes introduced
in R version 2.12.  However I haven't seen those used at PacBio yet and it's
less elegant, so will avoid introducing new syntax.
}

