% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pbsmrtpipe.R
\name{registerToolContract}
\alias{registerToolContract}
\title{Register a MetaTask}
\usage{
registerToolContract(toolContract, toCmd)
}
\arguments{
\item{taskId}{The metatask id}

\item{taskType}{the task type}

\item{inputTypes}{a list of file types}

\item{outputTypes}{a list of file types}

\item{taskOptions}{a hash of taskOptionId -> JsonSchema}

\item{nproc}{the number of processors}

\item{resources}{the list of Resource types}
}
\value{
A metaTask instance
}
\description{
Register a MetaTask
}

