% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pbsmrtpipe.R
\docType{data}
\name{ResourceTypes}
\alias{ResourceTypes}
\title{Symbol types used in Tool Contract and Resolved Tool Contracts}
\format{\preformatted{List of 3
 $ TMP_DIR : chr "$tmpdir"
 $ TMP_FILE: chr "$tmpfile"
 $ LOG_FILE: chr "$logfile"
}}
\usage{
ResourceTypes
}
\description{
These are log, tmp files, and dirs.
}
\keyword{datasets}

