% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/driver.R
\name{runDriver}
\alias{runDriver}
\title{Central Commandline Driver interface
Tools should use this interface to}
\usage{
runDriver(args_, toolContractRegistry_)
}
\description{
Central Commandline Driver interface
Tools should use this interface to
}

