% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\docType{data}
\name{SymbolTypes}
\alias{SymbolTypes}
\title{Symbol types used in Tool Contract and Resolved Tool Contracts}
\format{An object of class \code{list} of length 2.}
\usage{
SymbolTypes
}
\description{
max NPROC
}
\keyword{datasets}

