% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\docType{data}
\name{TaskTypes}
\alias{TaskTypes}
\title{PacBio Defined task types}
\format{An object of class \code{TaskTypes} of length 1.}
\usage{
TaskTypes
}
\description{
PacBio Defined task types
}
\keyword{datasets}

