% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/io.R
\name{writeToolContract}
\alias{writeToolContract}
\title{Write Tool Contract to JSON file}
\usage{
writeToolContract(toolContract, jsonPath)
}
\arguments{
\item{toolContract}{Tool Contract}
}
\description{
Write Tool Contract to JSON file
}

