% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/common.R
\docType{data}
\name{FileTypes}
\alias{FileTypes}
\title{Registry for all the PacBio defined file types}
\format{\preformatted{List of 5
 $ FASTA :Formal class 'FileType' [package "pbcommandR"] with 4 slots
  .. ..@ fileTypeId: chr "PacBio.FileTypes.Fasta"
  .. ..@ baseName  : chr "file"
  .. ..@ fileExt   : chr "fasta"
  .. ..@ mimeType  : chr "text/plain"
 $ FASTQ :Formal class 'FileType' [package "pbcommandR"] with 4 slots
  .. ..@ fileTypeId: chr "PacBio.FileTypes.Fastq"
  .. ..@ baseName  : chr "file"
  .. ..@ fileExt   : chr "fastq"
  .. ..@ mimeType  : chr "text/plain"
 $ GFF   :Formal class 'FileType' [package "pbcommandR"] with 4 slots
  .. ..@ fileTypeId: chr "PacBio.FileTypes.Gff"
  .. ..@ baseName  : chr "file"
  .. ..@ fileExt   : chr "gff"
  .. ..@ mimeType  : chr "text/plain"
 $ REPORT:Formal class 'FileType' [package "pbcommandR"] with 4 slots
  .. ..@ fileTypeId: chr "PacBio.FileTypes.JsonReport"
  .. ..@ baseName  : chr "file.report"
  .. ..@ fileExt   : chr "json"
  .. ..@ mimeType  : chr "application/json"
 $ TXT   :Formal class 'FileType' [package "pbcommandR"] with 4 slots
  .. ..@ fileTypeId: chr "PacBio.FileTypes.txt"
  .. ..@ baseName  : chr "file"
  .. ..@ fileExt   : chr "txt"
  .. ..@ mimeType  : chr "text/plain"
}}
\usage{
FileTypes
}
\description{
Registry for all the PacBio defined file types
}
\keyword{datasets}

