% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\docType{data}
\name{ResourceTypes}
\alias{ResourceTypes}
\title{Symbol types used in Tool Contract and Resolved Tool Contracts}
\format{An object of class \code{list} of length 3.}
\usage{
ResourceTypes
}
\description{
These are log, tmp files, and dirs.
}
\keyword{datasets}

