% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pbsmrtpipe.R
\docType{data}
\name{SymbolTypes}
\alias{SymbolTypes}
\title{Symbol types used in Tool Contract and Resolved Tool Contracts}
\format{\preformatted{List of 2
 $ MAX_NPROC  : chr "$max_nproc"
 $ MAX_NCHUNKS: chr "$max_nchunks"
}}
\usage{
SymbolTypes
}
\description{
max NPROC
}
\keyword{datasets}

