% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/common.R
\docType{data}
\name{TaskTypes}
\alias{TaskTypes}
\title{PacBio Defined task types}
\format{\preformatted{Formal class 'TaskTypes' [package "pbcommandR"] with 2 slots
  ..@ LOCAL      : chr "pbsmrtpipe.constants.local_task"
  ..@ DISTRIBUTED: chr "pbsmrtpipe.task_types.distributed_task"
}}
\usage{
TaskTypes
}
\description{
PacBio Defined task types
}
\keyword{datasets}

